move some functions into utils.py
